backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
requests==2.31.0
deep-translator==1.11.4
groq==0.4.1
python-dotenv==1.0.0
